chore: optimzie failer
