+++ support/convert/convert_test.go
@@ -164,29 +164,3 @@ func TestCopyBytes(t *testing.T) {
 		assert.Equal(t, 0, cap(copied))
 	})
 }
-
-func TestToAnySlice(t *testing.T) {
-	t.Run("empty slice", func(t *testing.T) {
-		var input []int
-		result := ToAnySlice(input...)
-		assert.NotNil(t, result)
-		assert.Empty(t, result)
-	})
-
-	t.Run("string slice", func(t *testing.T) {
-		input := []string{"a", "b", "c"}
-		assert.Equal(t, []any{"a", "b", "c"}, ToAnySlice(input...))
-	})
-
-	t.Run("int slice", func(t *testing.T) {
-		assert.Equal(t, []any{1, 2, 3}, ToAnySlice(1, 2, 3))
-	})
-
-	t.Run("bool slice", func(t *testing.T) {
-		assert.Equal(t, []any{true, false}, ToAnySlice(true, false))
-	})
-
-	t.Run("float slice", func(t *testing.T) {
-		assert.Equal(t, []any{1.1, 2.2, 3.3}, ToAnySlice(1.1, 2.2, 3.3))
-	})
-}
